requirements.txt
llama-index==0.11.17
llama-index-llms-openai==0.2.10
python-dotenv==1.0.1
jupyter==1.0.0